backoffice/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.8.0
setuptools==70.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
wheel==0.43.0
